backend/requirements.txt
MetaTrader5==5.0.45
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
psycopg2-binary==2.9.9
